import objects

This is inspired by the python idiom from module import object as new_name.

import_from(
  x,
  ...,
  .into = parent.frame(),
  .parent = .GlobalEnv,
  .chdir = FALSE,
  .recursive = FALSE,
  .pos = 2L
)

Arguments

x
a bare symbol name of a package, a character vector of filepaths, an environment (which could be a python module), or any object with names and [[ methods defined.

...
objects to import from x into .into. if named, the name will be the the new name after import. Alternatively, you can also supply the wildcard string "*" or "**", along with some additional overrides. See examples for details.

.into
An R environment, or something coercible to one by as.environment, or a character string that is the name of a (potentially new) attached environment. The default is the current frame.

.parent, .chdir, .recursive
Only applicable if x is a character vector of filepaths to R scripts, in which case these are passed on to include (chdir, recursive) or new.env(parent)

.pos
Only applicable if .into is a string that is the name of a new environment that will be attached, in which case this will be the position on new environment on the search path.

Value

the R environment or object that x resolved to, invisibly.

Note

If x is a package name, then no check is performed to ensure the object being imported is an exported function. As such, import_from() can be used to access package internal objects, though doing so is usually bad practice.

Examples

show_whats_imported <- function(...) {
  import_from(...)
  setdiff(names(environment()), "...")
}

## Importing from an R package
# import one object
show_whats_imported(envir, include)

#> [1] "include"

# rename an object on import
show_whats_imported(envir, sys_source = include)

#> [1] "sys_source"

# import all NAMESPACE exports
show_whats_imported(envir, "*")

#> [1] "include"       "import_from"   "attach_source" "attach_eval"  

show_whats_imported(envir) # missing `...` is interpreted as "*"

#> [1] "include"       "import_from"   "attach_source" "attach_eval"  

# import all NAMESPACE exports, except for `include`
show_whats_imported(envir, "*", -include)

#> [1] "import_from"   "attach_source" "attach_eval"  

# import all NAMESPACE exports, except rename `include` to `sys_source`
show_whats_imported(envir, "*", sys_source = include)

#> [1] "sys_source"    "import_from"   "attach_source" "attach_eval"  

# exclude more than one
show_whats_imported(envir, "*", -include, -attach_eval)

#> [1] "import_from"   "attach_source"

show_whats_imported(envir, "*", -c(include, attach_eval))

#> [1] "import_from"   "attach_source"

# import all NAMESPACE exports, also one internal function names `find_r_files`
show_whats_imported(envir, "*", find_r_files)

#> [1] "find_r_files"  "include"       "import_from"   "attach_source"
#> [5] "attach_eval"  

# import ALL package functions, including all internal functions
show_whats_imported(envir, "**")

#>  [1] "include"                  "import_from"             
#>  [3] "attach_source"            "complete_names"          
#>  [5] "set_library_default_pos"  "as_maybe_attached_env"   
#>  [7] "find_r_files"             "warn_about_conflicts"    
#>  [9] "within.environment"       "as_tidy_env_name"        
#> [11] "resolve_wildcard_imports" "within.character"        
#> [13] "attach_eval"             

# import ALL objects in the package NAMESPACE, including R's NAMESPACE machinery
show_whats_imported(envir, "***")

#>  [1] "include"                  "import_from"             
#>  [3] ".__S3MethodsTable__."     ".__NAMESPACE__."         
#>  [5] "attach_source"            "complete_names"          
#>  [7] "set_library_default_pos"  ".packageName"            
#>  [9] "as_maybe_attached_env"    "find_r_files"            
#> [11] "warn_about_conflicts"     "within.environment"      
#> [13] "as_tidy_env_name"         "resolve_wildcard_imports"
#> [15] "within.character"         "attach_eval"             

## Importing from R files
# setup
dir.create(tmpdir <- tempfile())
owd <- setwd(tmpdir)
writeLines(c("useful_function <- function() 'I am useful'",
             ".less_useful_fn <- function() 'less useful'"),
           "my_helpers.R")

# import one function by name
show_whats_imported("my_helpers.R", useful_function)

#> [1] "useful_function"

# import all objects whose names don't start with a "." or "_"
show_whats_imported("my_helpers.R", "*")

#> [1] "useful_function"

# import all objects
show_whats_imported("my_helpers.R", "**")

#> [1] "useful_function" ".less_useful_fn"

# if the filepath to your scripts is stored in a variable, supply it in a call
x <- "my_helpers.R"
try(show_whats_imported(x)) # errors out, because no package 'x'

#> Error in loadNamespace(name) : there is no package called ‘x’

# to force the value to be used, just supply it as a call rather than a bare symbol.
# the simplest call can be just wrapping in () or {}
show_whats_imported({x})

#> [1] "useful_function"

show_whats_imported((x))

#> [1] "useful_function"

show_whats_imported(c(x))

#> [1] "useful_function"

show_whats_imported({{x}}) # tidyverse style unquoting

#> [1] "useful_function"

## Importing R objects

# if you have an actual R object that you want to import from, you will
# have to supply it in a call
x <- list(obj1 = "one", obj2 = "two")
show_whats_imported({x})

#> [1] "obj2" "obj1"

# \dontrun{
  # don't run this so we don't take a reticulate dependency
  import_from(reticulate, py_module = import) # rename object on import

  # import one object
  show_whats_imported(py_module("numpy"), random)

#> [1] "random"

  # to prevent automatic conversion
  show_whats_imported(py_module("numpy", convert = FALSE), random)

#> [1] "random"

  # import all objects that don't begin with a `_`
  # by default, other modules found in the module are also not imported
  show_whats_imported(py_module("glob"), "*")

#> [1] "magic_check_bytes" "magic_check"       "iglob"            
#> [4] "has_magic"         "glob1"             "glob0"            
#> [7] "glob"              "escape"           

  # to import EVERYTHING pass "**"
  # now includes modules that your modules imported, like `os`
  show_whats_imported(py_module("glob"), "**")

#>  [1] "re"                "os"                "magic_check_bytes"
#>  [4] "magic_check"       "iglob"             "has_magic"        
#>  [7] "glob1"             "glob0"             "glob"             
#> [10] "fnmatch"           "escape"           

  rm(py_module) # clean up
# }

# cleanup
setwd(owd)
unlink(tmpdir, recursive = TRUE)
rm(show_whats_imported, tmpdir, owd)